Added User profile editing. Removed :bio added :display_name

+++ client/src/components/ProjectAddUsersForm.js
@@ -39,37 +39,45 @@ function ProjectAddUsersForm(props) {
   }
 
   function removeUser(user) {
-    fetch(`/projects/${project.id}/users/${user.id}`, {
-      method: 'DELETE'
-    }).then(res => {
-      if (res.ok) {
-        res.json().then(data => {
-          updateProjectCallback(data);
-          setErrors([]);
-        });
-      } else {
-        res.json().then(data => setErrors(data.errors));
-      }
-    }).catch(err => {
-      console.log(err);
-    });
+    if (window.confirm(`Are you sure you want to remove ${user.display_name || user.username} from the project?`)) {
+      fetch(`/projects/${project.id}/users/${user.id}`, {
+        method: 'DELETE'
+      }).then(res => {
+        if (res.ok) {
+          res.json().then(data => {
+            updateProjectCallback(data);
+            setErrors([]);
+          });
+        } else {
+          res.json().then(data => setErrors(data.errors));
+        }
+      }).catch(err => {
+        console.log(err);
+      });
+    }
   }
 
   function setOwner(user, bool) {
-    fetch(`/projects/${project.id}/users/${user.id}/${bool ? 'owner' : 'not_owner'}`, {
-      method: 'PATCH'
-    }).then(res => {
-      if (res.ok) {
-        res.json().then(data => {
-          updateProjectCallback(data);
+    if (window.confirm(`Are you sure you want to change ${user.display_name || user.username}'s permissions?`)) {
+      fetch(`/projects/${project.id}/users/${user.id}/${bool ? 'owner' : 'not_owner'}`, {
+        method: 'PATCH'
+      }).then(res => {
+        if (res.ok) {
+          res.json().then(data => {
+            updateProjectCallback(data);
+          });
+        } else {
+          res.json().then(data => {
+            console.log(data);
+            setErrors(data.errors);
+          });
+        }
+      })
+        .catch(err => {
+          console.log(err);
+          setErrors(err.errors);
         });
-      } else {
-        res.json().then(data => {
-          console.log(data);
-          setErrors(data.errors);
-        });
-      }
-    });
+    }
   }
 
   return (
@@ -97,9 +105,9 @@ function ProjectAddUsersForm(props) {
               <div style={{ display: "flex", alignItems: "center", justifyContent: "space-between", width: "100%" }}>
                 {project.owners.map(u => u.id).includes(user.id) ? (
                   <>
-                    <span>
+                    <span style={{ display: "flex", alignItems: "center" }}>
                       <OwnerAvatar src={user.image_url} alt="" />
-                      <strong>{user.username}</strong> (Owner)
+                      <strong>{user.display_name || user.username}</strong> (Owner)
                     </span>
                     <span>
                       <InlineButton onClick={() => setOwner(user, false)}><span style={{ color: "red" }}>&times;</span> Owner</InlineButton>
@@ -107,9 +115,9 @@ function ProjectAddUsersForm(props) {
                   </>
                 ) : (
                   <>
-                    <span>
+                    <span style={{ display: "flex", alignItems: "center" }}>
                       <OwnerAvatar src={user.image_url} alt="" />
-                      <strong>{user.username}</strong>
+                      <strong>{user.display_name || user.username}</strong>
                     </span>
                     <span>
                       <InlineButton style={{ color: "red" }} onClick={() => removeUser(user)}><i className="fa fa-trash-alt"></i></InlineButton>
@@ -137,8 +145,8 @@ const UsersList = styled.ul`
 
 const OwnerAvatar = styled.img`
   display: inline;
-  width: 30px;
-  height: 30px;
+  width: 25px;
+  height: 25px;
   border-radius: 50%;
   margin-right: 5px;
 `;
@@ -146,6 +154,7 @@ const OwnerAvatar = styled.img`
 const InlineButton = styled.button`
   background: none;
   margin: 0 0.25rem;
+  padding: 2px 5px;
   `;
